fix some code style
