--- sharding-opentracing/.travis.yml
@@ -1,7 +1,0 @@
-language: java
-jdk:
-  - oraclejdk8
-before_script:
-  - echo "MAVEN_OPTS='-Xmx1024m -XX:MaxPermSize=256m'" > ~/.mavenrc
-after_success:
-  - mvn clean cobertura:cobertura coveralls:report
