+++ sharding-opentracing/src/test/java/io/shardingjdbc/opentracing/ExecuteEventListenerTest.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -54,7 +54,7 @@ public class ExecuteEventListenerTest {
     }
     
     @Before
-    public void before() throws NoSuchFieldException, IllegalAccessException {
+    public void before() {
         TRACER.reset();
     }
