+++ sharding-opentracing/src/test/java/io/shardingjdbc/opentracing/fixture/FooTracer.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -31,9 +31,7 @@ public class FooTracer implements Tracer {
     }
     
     @Override
-    public <C> void inject(final SpanContext spanContext, final Format<C> format, final C c) {
-        
-    }
+    public <C> void inject(final SpanContext spanContext, final Format<C> format, final C c) {}
     
     @Override
     public <C> SpanContext extract(final Format<C> format, final C c) {
